+++ packages/apollo-link-http/src/httpLink.ts
@@ -31,7 +31,9 @@ const parseAndCheckResponse = request => (response: Response) => {
     })
     .catch(e => {
       const httpError = new Error(
-        `Network request failed with status ${response.status} - "${response.statusText}"`,
+        `Network request failed with status ${response.status} - "${
+          response.statusText
+        }"`,
       ) as ResponseError;
       httpError.response = response;
       httpError.parseError = e;
@@ -92,6 +94,12 @@ export interface FetchOptions {
   headers?: any;
   fetchOptions?: any;
 }
+
+const defaultHttpOptions = {
+  includeQuery: true,
+  includeExtensions: false,
+};
+
 export const createHttpLink = (
   {
     uri,
@@ -116,15 +124,17 @@ export const createHttpLink = (
           credentials,
           fetchOptions = {},
           uri: contextURI,
+          http: httpOptions = {},
         } = operation.getContext();
         const { operationName, extensions, variables, query } = operation;
+        const http = { ...defaultHttpOptions, ...httpOptions };
+        const body = { operationName, variables };
 
-        const body = {
-          operationName,
-          variables,
-          query: print(query),
-        };
-        if (includeExtensions) (body as any).extensions = extensions;
+        if (includeExtensions || http.includeExtensions)
+          (body as any).extensions = extensions;
+
+        // not sending the query (i.e persisted queries)
+        if (http.includeQuery) body.query = print(query);
 
         let serializedBody;
         try {